refactor(decode): use declarative/reusable impl
These sensors have several things in common, and the parsing
algorithms differ mostly just by the data's location in the
incoming message.

I worked on this while waiting for a battery misread from an
H5075 because somehow it gets battery readings of >200. The
plan is to implement another battery parser once data for the
bad reads is available.